+++ GeosCore/global_ch4_mod.F90
@@ -23,7 +23,6 @@ MODULE GLOBAL_CH4_MOD
 !
 ! !PUBLIC MEMBER FUNCTIONS:
 !
-  PUBLIC :: EMISSCH4
   PUBLIC :: CHEMCH4
   PUBLIC :: INIT_GLOBAL_CH4
 !
@@ -55,519 +54,6 @@ MODULE GLOBAL_CH4_MOD
   REAL(fp)              :: TROPOCH4
 
 CONTAINS
-!EOC
-!------------------------------------------------------------------------------
-!                  GEOS-Chem Global Chemical Transport Model                  !
-!------------------------------------------------------------------------------
-!BOP
-!
-! !IROUTINE: emissch4
-!
-! !DESCRIPTION: Subroutine EMISSCH4 places emissions of CH4 [kg] into the
-!  chemical species array.
-!\\
-!\\
-! !INTERFACE:
-!
-  SUBROUTINE EMISSCH4( Input_Opt, State_Chm, State_Grid, State_Met, RC )
-!
-! !USES:
-!
-    USE HCO_Utilities_GC_Mod, ONLY : HCO_GC_EvalFld
-    USE HCO_Utilities_GC_Mod, ONLY : HCO_GC_GetDiagn
-    USE ErrCode_Mod
-    USE Input_Opt_Mod,        ONLY : OptInput
-    USE State_Chm_Mod,        ONLY : ChmState
-    USE State_Met_Mod,        ONLY : MetState
-    USE State_Grid_Mod,       ONLY : GrdState
-!
-! !INPUT PARAMETERS:
-!
-    TYPE(OptInput), INTENT(IN)    :: Input_Opt   ! Input Options object
-    TYPE(ChmState), INTENT(INOUT) :: State_Chm   ! Chemistry State object
-    TYPE(GrdState), INTENT(IN)    :: State_Grid  ! Grid State object
-    TYPE(MetState), INTENT(IN)    :: State_Met   ! Meteorology State object
-!
-! !OUTPUT PARAMETERS:
-!
-    INTEGER,        INTENT(OUT)   :: RC          ! Success or failure?
-!
-! !REMARKS:
-!  WARNING: Soil absorption has to be the 15th field in CH4_EMIS
-!  Also: the ND58 diagnostics have now been removed.  We still need to
-!  read the HEMCO manual diagnostics into CH4_EMIS for the analytical
-!  inversion.  Therefore, we will keep EmissCh4 for the time-being
-!  but only remove the bpch diagnostic.
-!
-! !REVISION HISTORY:
-!  (1 ) Created by Bryan Duncan (1/99).  Adapted for CH4 chemistry by
-!        James Wang (7/00).  Inserted into module "global_ch4_mod.f"
-!        by Bob Yantosca. (bmy, 1/16/01)
-!  See https://github.com/geoschem/geos-chem for complete history
-!EOP
-!------------------------------------------------------------------------------
-!BOC
-!
-! !LOCAL VARIABLES:
-!
-    ! Scalars
-    INTEGER            :: I, J, N
-
-    ! Strings
-    CHARACTER(LEN= 63) :: DgnName
-    CHARACTER(LEN=255) :: ErrMsg
-    CHARACTER(LEN=255) :: ThisLoc
-
-    ! Logicals
-    LOGICAL, SAVE      :: FIRST = .TRUE.
-
-    ! Arrays of state vector elements for applying emissions perturbations
-    REAL(fp)           :: STATE_VECTOR(State_Grid%NX,State_Grid%NY)
-
-    ! Array of scale factors for emissions (from HEMCO)
-    REAL(fp)           :: EMIS_SF(State_Grid%NX,State_Grid%NY)
-
-    ! Pointers
-    REAL(f4), POINTER  :: Ptr2D(:,:)
-
-    !=================================================================
-    ! EMISSCH4 begins here!
-    !=================================================================
-
-    ! Nullify pointers
-    Ptr2D => NULL()
-
-    ! Assume success
-    RC      = GC_SUCCESS
-    ErrMsg  = ''
-    ThisLoc = ' -> at EMISSCH4 (in GeosCore/global_ch4_mod.F90)'
-
-    IF ( Input_Opt%Verbose ) THEN
-       print*,'BEGIN SUBROUTINE: EMISSCH4'
-    ENDIF
-
-    ! =================================================================
-    ! Get fields for CH4 analytical inversions if needed
-    ! =================================================================
-    IF ( Input_Opt%DoAnalyticalInv ) THEN
-
-       ! Evaluate the state vector field from HEMCO
-       CALL HCO_GC_EvalFld( Input_Opt, State_Grid, 'CH4_STATE_VECTOR', &
-                         STATE_VECTOR, RC)
-       IF ( RC /= GC_SUCCESS ) THEN
-          ErrMsg = 'CH4_STATE_VECTOR not found in HEMCO data list!'
-          CALL GC_Error( ErrMsg, RC, ThisLoc )
-          RETURN
-       ENDIF
-
-    ENDIF
-
-    IF ( Input_Opt%UseEmisSF ) THEN
-
-       ! Evaluate CH4 emissions scale factors from HEMCO
-       CALL HCO_GC_EvalFld( Input_Opt, State_Grid, 'EMIS_SF', EMIS_SF, RC)
-       IF ( RC /= GC_SUCCESS ) THEN
-          ErrMsg = 'EMIS_SF not found in HEMCO data list!'
-          CALL GC_Error( ErrMsg, RC, ThisLoc )
-          RETURN
-       ENDIF
-
-    ENDIF
-
-    ! =================================================================
-    ! --> All emission calculations are now done through HEMCO
-    ! HEMCO stores emissions of all species internally in the HEMCO
-    ! state object. Here, we pass these emissions into module array
-    ! CH4_EMIS in units kg/m2/s. These values are then either added to
-    ! the species array (full mixing scheme) or used later on in
-    ! vdiff_mod.F90 if the non-local PBL mixing scheme is used.
-    !
-    ! The CH4_EMIS array is mostly used for backwards compatibility
-    ! (especially the diagnostics). It is also used to ensure that
-    ! in a multi-species simulation, species 1 (total CH4) is properly
-    ! defined.
-    !
-    !                                              (ckeller, 9/12/2013)
-    ! =================================================================
-    State_Chm%CH4_EMIS(:,:,:) = 0e+0_fp
-
-    !-------------------
-    ! Oil
-    !-------------------
-    DgnName = 'CH4_OIL'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,2) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Gas
-    !-------------------
-    DgnName = 'CH4_GAS'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,3) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Coal
-    !-------------------
-    DgnName = 'CH4_COAL'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,4) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Livestock
-    !-------------------
-    DgnName = 'CH4_LIVESTOCK'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,5) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Landfills
-    !-------------------
-    DgnName = 'CH4_LANDFILLS'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,6) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Wastewater
-    !-------------------
-    DgnName = 'CH4_WASTEWATER'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,7) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Rice
-    !-------------------
-    DgnName = 'CH4_RICE'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,8) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Other anthropogenic
-    !-------------------
-    DgnName = 'CH4_ANTHROTHER'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc = ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,9) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Biomass burning
-    !-------------------
-    DgnName = 'CH4_BIOMASS'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,10) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Wetland
-    !-------------------
-    DgnName = 'CH4_WETLAND'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,11) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Global seeps
-    !-------------------
-    DgnName = 'CH4_SEEPS'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,12) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Lakes
-    !-------------------
-    DgnName = 'CH4_LAKES'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,13) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Termites
-    !-------------------
-    DgnName = 'CH4_TERMITES'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,14) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Soil absorption (those are negative!)
-    !-------------------
-    DgnName = 'CH4_SOILABSORB'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,15) =  Ptr2D(:,:) * -1.0_fp
-    ENDIF
-    Ptr2D => NULL()
-
-    !-------------------
-    ! Reservoirs
-    !-------------------
-    DgnName = 'CH4_RESERVOIRS'
-    CALL HCO_GC_GetDiagn( Input_Opt, State_Grid, DgnName, .FALSE., RC, Ptr2D=Ptr2D )
-
-    ! Trap potential errors and assign HEMCO pointer to array
-    IF ( RC /= GC_SUCCESS ) THEN
-       ErrMsg = 'Cannot get pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Error( ErrMsg, RC, ThisLoc )
-       RETURN
-    ELSEIF ( .NOT. ASSOCIATED(Ptr2D) ) THEN
-       ErrMsg = 'Unassociated pointer to HEMCO field ' // TRIM(DgnName)
-       CALL GC_Warning( ErrMsg, RC, ThisLoc=ThisLoc )
-    ELSE
-       State_Chm%CH4_EMIS(:,:,16) =  Ptr2D(:,:)
-    ENDIF
-    Ptr2D => NULL()
-
-    ! =================================================================
-    ! Total emission: sum of all emissions - (2*soil absorption)
-    ! We have to substract soil absorption twice because it is added
-    ! to other emissions in the SUM function. (ccc, 7/23/09)
-    ! =================================================================
-    State_Chm%CH4_EMIS(:,:,1) = SUM(State_Chm%CH4_EMIS, 3) - (2 * State_Chm%CH4_EMIS(:,:,15))
-
-    IF ( Input_Opt%Verbose ) THEN
-       WRITE(*,*) 'CH4_EMIS (kg/m2/s):'
-       WRITE(*,*) 'Total        : ', SUM(State_Chm%CH4_EMIS(:,:,1))
-       WRITE(*,*) 'Oil          : ', SUM(State_Chm%CH4_EMIS(:,:,2))
-       WRITE(*,*) 'Gas          : ', SUM(State_Chm%CH4_EMIS(:,:,3))
-       WRITE(*,*) 'Coal         : ', SUM(State_Chm%CH4_EMIS(:,:,4))
-       WRITE(*,*) 'Livestock    : ', SUM(State_Chm%CH4_EMIS(:,:,5))
-       WRITE(*,*) 'Landfills    : ', SUM(State_Chm%CH4_EMIS(:,:,6))
-       WRITE(*,*) 'Wastewater   : ', SUM(State_Chm%CH4_EMIS(:,:,7))
-       WRITE(*,*) 'Rice         : ', SUM(State_Chm%CH4_EMIS(:,:,8))
-       WRITE(*,*) 'Other anth   : ', SUM(State_Chm%CH4_EMIS(:,:,9))
-       WRITE(*,*) 'Biomass burn : ', SUM(State_Chm%CH4_EMIS(:,:,10))
-       WRITE(*,*) 'Wetlands     : ', SUM(State_Chm%CH4_EMIS(:,:,11))
-       WRITE(*,*) 'Seeps        : ', SUM(State_Chm%CH4_EMIS(:,:,12))
-       WRITE(*,*) 'Lakes        : ', SUM(State_Chm%CH4_EMIS(:,:,13))
-       WRITE(*,*) 'Termites     : ', SUM(State_Chm%CH4_EMIS(:,:,14))
-       WRITE(*,*) 'Soil absorb  : ', SUM(State_Chm%CH4_EMIS(:,:,15))
-       WRITE(*,*) 'Reservoirs   : ', SUM(State_Chm%CH4_EMIS(:,:,16))
-    ENDIF
-
-    ! =================================================================
-    ! Do scaling for analytical inversion
-    ! =================================================================
-    IF ( Input_Opt%DoAnalyticalInv  .or. &
-         Input_Opt%UseEmisSF      .or. &
-         Input_Opt%UseOHSF        ) THEN
-
-       ! Don't optimize for soil absorption so remove from the total
-       ! emissions array
-       State_Chm%CH4_EMIS(:,:,1) = State_Chm%CH4_EMIS(:,:,1) + State_Chm%CH4_EMIS(:,:,15)
-
-       !$OMP PARALLEL DO       &
-       !$OMP DEFAULT( SHARED ) &
-       !$OMP PRIVATE( I, J)	 
-       DO J = 1, State_Grid%NY
-       DO I = 1, State_Grid%NX
-
-          !------------------------------------------------------------
-          ! Apply emission scale factors from a previous inversion
-          !------------------------------------------------------------
-          IF ( Input_Opt%UseEmisSF ) THEN
-             ! Scale total emissions
-             State_Chm%CH4_EMIS(I,J,1) = State_Chm%CH4_EMIS(I,J,1) * EMIS_SF(I,J)
-          ENDIF
-
-          !------------------------------------------------------------
-          ! Perturb emissions for analytical inversion
-          !------------------------------------------------------------
-          IF ( Input_Opt%DoAnalyticalInv ) THEN
-
-             ! Only apply emission perturbation to current state vector
-             ! element number
-             IF ( Input_Opt%StateVectorElement .GT. 0 ) THEN
-
-                ! Convert STATE_VECTOR value to nearest integer for comparison
-                IF ( NINT(STATE_VECTOR(I,J)) ==                              &
-                     Input_Opt%StateVectorElement) THEN
-                   State_Chm%CH4_EMIS(I,J,1) = State_Chm%CH4_EMIS(I,J,1)     &
-                                             * Input_Opt%EmisPerturbFactor
-
-                   IF ( Input_Opt%Verbose ) THEN
-                      Print*, 'Analytical Inversion: ',                      &
-                              'Scaled state vector element ',                &
-                              Input_Opt%StateVectorElement, ' by ',          &
-                              Input_Opt%EmisPerturbFactor
-                   ENDIF
-                ENDIF
-             ENDIF
-          ENDIF
-
-       ENDDO
-       ENDDO
-       !$OMP END PARALLEL DO
-
-       ! Now that we've done the emission factor scaling, add soil absorption
-       ! back to the total emissions array
-       State_Chm%CH4_EMIS(:,:,1) = State_Chm%CH4_EMIS(:,:,1) - State_Chm%CH4_EMIS(:,:, 15)
-
-    ENDIF
-
-    IF ( Input_Opt%Verbose ) THEN
-       print*,'END SUBROUTINE: EMISSCH4'
-    ENDIF
-
-  END SUBROUTINE EMISSCH4
 !EOC
 !------------------------------------------------------------------------------
 !                  GEOS-Chem Global Chemical Transport Model                  !
@@ -848,8 +334,8 @@ CONTAINS
     ! Pointers
     TYPE(SpcConc), POINTER :: Spc(:)
 
-    ! Array of scale factors for OH (from HEMCO)
-    REAL(fp)           :: OH_SF(State_Grid%NX,State_Grid%NY)
+!    ! Array of scale factors for OH (from HEMCO)
+!    REAL(fp)           :: OH_SF(State_Grid%NX,State_Grid%NY)
 
     !=================================================================
     ! CH4_DECAY begins here!
@@ -864,20 +350,20 @@ CONTAINS
     ! Point to the chemical species array
     Spc => State_Chm%Species
 
-    ! =================================================================
-    ! Get fields for CH4 analytical inversions if needed
-    ! =================================================================
-    IF ( Input_Opt%UseOHSF ) THEN
-
-       ! Evaluate OH scale factors from HEMCO
-       CALL HCO_GC_EvalFld( Input_Opt, State_Grid, 'OH_SF', OH_SF, RC)
-       IF ( RC /= GC_SUCCESS ) THEN
-          ErrMsg = 'OH_SF not found in HEMCO data list!'
-          CALL GC_Error( ErrMsg, RC, ThisLoc )
-          RETURN
-       ENDIF
-
-    ENDIF
+!    ! =================================================================
+!    ! Get fields for CH4 analytical inversions if needed
+!    ! =================================================================
+!    IF ( Input_Opt%UseOHSF ) THEN
+!
+!       ! Evaluate OH scale factors from HEMCO
+!       CALL HCO_GC_EvalFld( Input_Opt, State_Grid, 'OH_SF', OH_SF, RC)
+!       IF ( RC /= GC_SUCCESS ) THEN
+!          ErrMsg = 'OH_SF not found in HEMCO data list!'
+!          CALL GC_Error( ErrMsg, RC, ThisLoc )
+!          RETURN
+!       ENDIF
+!
+!    ENDIF
 
     !=================================================================
     ! %%%%% HISTORY (aka netCDF diagnostics) %%%%%
@@ -937,14 +423,14 @@ CONTAINS
           ! BOH from HEMCO in units of kg/m3, convert to molec/cm3
           C_OH = State_Chm%BOH(I,J,L) * XNUMOL_OH / CM3PERM3
 
-          ! Apply OH scale factors from a previous inversion
-          IF ( Input_Opt%UseOHSF ) THEN
-             C_OH = C_OH * OH_SF(I,J)
-             IF ( Input_Opt%Verbose ) THEN
-                !This will print over every grid box; comment out for now
-                !Print*, 'Applying scale factor to OH: ', OH_SF(I,J)
-             ENDIF
-          ENDIF
+!          ! Apply OH scale factors from a previous inversion
+!          IF ( Input_Opt%UseOHSF ) THEN
+!             C_OH = C_OH * OH_SF(I,J)
+!             IF ( Input_Opt%Verbose ) THEN
+!                !This will print over every grid box; comment out for now
+!                !Print*, 'Applying scale factor to OH: ', OH_SF(I,J)
+!             ENDIF
+!          ENDIF
 
           ! Cl in [molec/cm3]
           ! BCl from HEMCO in units of mol/mol, convert to molec/cm3
@@ -1213,16 +699,6 @@ CONTAINS
              Ktrop = 1.64e-12_f8 * EXP( -1520.0_f8 / State_Met%T(I,J,L) )
              LossOHbyMCF = LossOHbyMCF + ( Ktrop * OHconc_MCM3 * airMass_m )
 
-             !---------------------------------------------------------------
-             ! HISTORY (aka netCDF diagnostics)
-             !
-             ! Keep track of CH4 emisisons [kg/s] for computing
-             ! the various lifetime metrics in post-processing
-             !---------------------------------------------------------------
-             IF ( L == 1 .and. State_Diag%Archive_CH4emission ) THEN
-                State_Diag%CH4emission(I,J) = State_Chm%CH4_EMIS(I,J,id_CH4)           &
-                                            * State_Grid%Area_M2(I,J)
-             ENDIF
           ENDIF
        ENDDO
 
@@ -1447,7 +923,7 @@ CONTAINS
 ! !IROUTINE: ch4_distrib
 !
 ! !DESCRIPTION: Subroutine CH4\_DISTRIB allocates the chemistry sink to
-!  different emission species. (ccc, 10/2/09)
+!  different CH4 species. (ccc, 10/2/09)
 !\\
 !\\
 ! !INTERFACE: